// client/src/app/interface/Resident.ts
export interface Resident {
    _id: string;
    first_name: string;
    middle_name: string;
    last_name: string;
    birth_date: string;
    address: string;
    civil_status: string;
    email: string;
    occupation: string;
    sex: string;
    phone_number: string;
    imageFile: string;
 }

 